+++ Scripts/Player/Sword.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public class Sword : MonoBehaviour 
 {
-    [SerializeField] private int _damage = 40;
+    [SerializeField] private float _damage = 40;
 
-    public int Damage { get; private set; }
+    public float Damage { get; private set; }
 
     private void Start()
     {
